Rename Server to StreamingServer and created a buffering Server facade around StreamingServer
